refactor(cli): move files that add options to commands into their own folder
